Add form component alignment tests
