+++ packages/dnb-design-system-portal/src/docs/uilib/layout/Examples.tsx
@@ -3,15 +3,26 @@ import styled from '@emotion/styled'
 import ComponentBox from '../../../shared/tags/ComponentBox'
 import MediaQuery from '@dnb/eufemia/src/shared/MediaQuery'
 import {
-  Grid,
-  Code,
-  Button,
-  Card,
-  Flex,
+  Input,
   H2,
+  Button,
+  ToggleButton,
+  Flex,
+  Switch,
+  Textarea,
+  Slider,
+  Radio,
+  Checkbox,
+  IconPrimary,
+  DatePicker,
+  Autocomplete,
+  Dropdown,
+  Space,
+  Card,
+  Code,
+  Grid,
   FormSet,
   FormRow,
-  Input,
 } from '@dnb/eufemia/src'
 import {
   TestElement,
@@ -22,6 +33,8 @@ import {
 import { defaultBreakpoints } from '@dnb/eufemia/src/shared/MediaQueryUtils'
 import { defaultQueries } from '@dnb/eufemia/src/shared/useMedia'
 import { Provider, useMedia, useMediaQuery } from '@dnb/eufemia/src/shared'
+import { SpacingElementProps } from '@dnb/eufemia/src/shared/types'
+import { ScrollView } from '@dnb/eufemia/src/fragments'
 
 export const LayoutComponents = () => {
   return (
@@ -304,6 +317,7 @@ export const ResponsiveGridContainer = () => {
   )
 }
 
+// Deprecated – can be removed in v11
 export const FormSetAlternativeBefore = () => (
   <ComponentBox>
     <FormSet label_direction="vertical">
@@ -346,4 +360,139 @@ export const FormSetAlternativeForms = () => (
       </Flex.Stack>
     </Form.Handler>
   </ComponentBox>
+)
+
+const AllComponents = ({
+  direction = 'vertical',
+  showText = false,
+  hideLabel = false,
+} = {}) => {
+  const params: SpacingElementProps = {
+    left: direction === 'horizontal' ? 'small' : null,
+    top: direction !== 'horizontal' ? 'small' : null,
+  }
+
+  let labels: Record<string, string> = {
+    datePicker: 'DatePicker:',
+    dropdown: 'Dropdown:',
+    autocomplete: 'Autocomplete:',
+    checkbox: 'Checkbox',
+    radio: 'Radio',
+    radioGroup: 'Radio Group:',
+    toggleButton: 'Toggle:',
+    toggleButtonGroup: 'Toggle Group:',
+    switch: 'Switch',
+    input: 'Input:',
+    textarea: 'Textarea:',
+    slider: 'Slider:',
+  }
+
+  if (hideLabel) {
+    labels = Object.entries(labels).reduce((acc, [k]) => {
+      acc[k] = ''
+      return acc
+    }, {})
+  }
+
+  const Components = () => (
+    <>
+      {showText && (
+        <>
+          <Space {...params} inline>
+            <p className="dnb-p">
+              paragraph{' '}
+              <IconPrimary
+                icon="bell"
+                size="medium"
+                {...params}
+                style={{ margin: 0 }} // since this is not a block element
+              />
+            </p>
+          </Space>
+          text
+        </>
+      )}
+      <Button text="Button" {...params} />
+      <Button icon="add" {...params} />
+      <Input label={labels.input} {...params} />
+      <Dropdown
+        label={labels.dropdown}
+        data={['Item A', 'Item B', 'Item C']}
+        {...params}
+      />
+      <Autocomplete
+        label={labels.autocomplete}
+        data={['Item A', 'Item B', 'Item C']}
+        {...params}
+      />
+      <DatePicker label={labels.datePicker} {...params} />
+      <IconPrimary icon="bell" size="medium" {...params} />
+      <Checkbox label={labels.checkbox} {...params} />
+      <Radio label={labels.radio} {...params} />
+      <Radio.Group label={labels.radioGroup} {...params}>
+        <Radio label={labels.radio} value="a" />
+        <Radio label={labels.radio} value="b" />
+      </Radio.Group>
+      <ToggleButton
+        label={labels.toggleButton}
+        text="Toggle"
+        {...params}
+      />
+      <ToggleButton.Group label={labels.toggleButtonGroup} {...params}>
+        <ToggleButton text="Toggle A" value="a" />
+        <ToggleButton text="Toggle B" value="b" />
+      </ToggleButton.Group>
+      <Switch label={labels.switch} {...params} />
+      <Textarea label={labels.textarea} rows="5" {...params} />
+      <div style={{ display: 'inline-flex' }}>
+        <Slider label={labels.slider} value={50} {...params} />
+      </div>
+    </>
+  )
+
+  if (direction === 'horizontal') {
+    return (
+      <Flex.Horizontal style={{ padding: '1rem', whiteSpace: 'nowrap' }}>
+        <Components />
+      </Flex.Horizontal>
+    )
+  } else {
+    Components._supportsSpacingProps = true
+    return (
+      <Flex.Vertical style={{ padding: '1rem' }}>
+        <Components />
+      </Flex.Vertical>
+    )
+  }
+}
+
+export const AllComponentsVerticalTestCase = (
+  props?: React.HTMLProps<HTMLDivElement>,
+) => (
+  <div data-visual-test="form-components-alignment-vertical" {...props}>
+    <AllComponents direction="vertical" />
+  </div>
+)
+
+export const AllComponentsVerticalLabelsTestCase = (
+  props?: React.HTMLProps<HTMLDivElement>,
+) => (
+  <div
+    data-visual-test="form-components-alignment-vertical-labels"
+    {...props}
+  >
+    <Provider formElement={{ label_direction: 'vertical' }}>
+      <AllComponents direction="vertical" />
+    </Provider>
+  </div>
+)
+
+export const AllComponentsHorizontalTestCase = (
+  props?: React.HTMLProps<HTMLDivElement>,
+) => (
+  <div data-visual-test="form-components-alignment-horizontal" {...props}>
+    <ScrollView>
+      <AllComponents direction="horizontal" />
+    </ScrollView>
+  </div>
 )
