+++ packages/dnb-design-system-portal/src/docs/uilib/layout/visual-tests.mdx
@@ -1,0 +1,5 @@
+import * as Examples from './Examples'
+
+<Examples.AllComponentsVerticalTestCase />
+<Examples.AllComponentsVerticalLabelsTestCase />
+<Examples.AllComponentsHorizontalTestCase />
